complet jwt-auth process

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
-const Goal = require("./../models/goalModel");
+const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc    Get goals
 // @route   GET/api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -22,6 +23,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(201).json(goal);
@@ -32,15 +34,27 @@ const setGoal = asyncHandler(async (req, res) => {
 // @access  Private
 const updateGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
-
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found .. !");
   }
 
+  // check if user exists
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found ... !");
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized .. !");
+  }
+
   const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
-  }); // 3th para is optional and use it if you want to create new one if the doesn exists
+  }); // 3th param is optional and use it if you want to create new one if the doesn exists
 
   res.status(200).json(updateGoal);
 });
@@ -50,14 +64,26 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @access  Private
 const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
-
   if (!goal) {
     res.status(400);
     throw new Error("Goal Not Found .. !");
   }
 
+  // check if user exists
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found ... !");
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized .. !");
+  }
+
   await goal.delete(); // also can use remove()
-  
+
   res.status(200).json({
     message: `delete successfully .. !`,
   });
